settings at home moved to a different widget
